Swap out Markdown -> MDX
Use actual MDX instead of just Markdown.
