+++ package.json
@@ -11,13 +11,13 @@
     "date-fns": "^2.24.0",
     "html-react-parser": "^1.3.0",
     "next": "^12.0.0",
+    "next-mdx-remote": "^3.0.8",
     "parse-link-header": "^1.0.1",
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-display-name": "^0.2.5",
     "react-dom": "^17.0.1",
     "react-icons": "^4.2.0",
-    "react-markdown": "^7.1.0",
     "react-oembed-container": "^1.0.0",
     "react-spring": "^9.2.4",
     "rooks": "^5.7.1",
