+++ src/@strapi/page/StrapiPage.js
@@ -1,20 +1,26 @@
+import PropTypes from 'prop-types';
 import { Article } from '@ui/box';
 import { Mdx } from '@ui/mdx';
 import { getPageSEOProps, PageSEO } from './seo';
 
-const PageBody = ({ body }) => {
+const PageBody = ({ source }) => {
   return (
     <Article>
-      <Mdx content={body} />
+      <Mdx source={source} />
     </Article>
   );
 };
+
+PageBody.propTypes = {
+  source: PropTypes.object.isRequired,
+};
+
 const templates = {
   full_screen: PageBody,
   page: PageBody,
 };
 
-export const StrapiPage = ({ page }) => {
+export const StrapiPage = ({ page, source }) => {
   const Component = templates[page.template];
 
   if (!Component) {
@@ -24,7 +30,7 @@ export const StrapiPage = ({ page }) => {
   return (
     <>
       <PageSEO {...getPageSEOProps(page)} />
-      <Component {...page} />
+      <Component {...page} source={source} />
     </>
   );
 };
